refoctor wrting task
