+++ reductstore/src/api/entry/write_single.rs
@@ -83,7 +83,8 @@ pub(crate) async fn write_record(
     };
 
     match check_request_and_get_sender.await {
-        Ok((ts, labels, tx)) => {
+        Ok((ts, labels, mut writer)) => {
+            let mut tx = writer.tx();
             macro_rules! send_chunk {
                 ($chunk:expr) => {
                     tx.send_timeout($chunk, IO_OPERATION_TIMEOUT)
