+++ reductstore/src/replication/replication_sender.rs
@@ -401,7 +401,7 @@ mod tests {
             .await
             .unwrap();
 
-        let tx = sender
+        let mut writer = sender
             .storage
             .write()
             .await
@@ -418,7 +418,11 @@ mod tests {
             sender.run().await
         });
 
-        tx.send(Ok(Some(Bytes::from("xxxx")))).await.unwrap();
+        writer
+            .tx()
+            .send(Ok(Some(Bytes::from("xxxx"))))
+            .await
+            .unwrap();
         sleep(Duration::from_millis(100)).await;
 
         let diagnostics = hourly_diagnostics.read().await.diagnostics();
@@ -594,7 +598,7 @@ mod tests {
             Err(_err) => storage.get_bucket_mut("src").unwrap(),
         };
 
-        let tx = bucket
+        let mut writer = bucket
             .write_record(
                 "test",
                 transaction.timestamp().clone(),
@@ -604,13 +608,15 @@ mod tests {
             )
             .await
             .unwrap();
-        tx.send(Ok(Some(Bytes::from(
-            (0..size).map(|_| 'x').collect::<String>(),
-        ))))
-        .await
-        .unwrap();
-        tx.send(Ok(None)).await.unwrap();
-        tx.closed().await;
+        writer
+            .tx()
+            .send(Ok(Some(Bytes::from(
+                (0..size).map(|_| 'x').collect::<String>(),
+            ))))
+            .await
+            .unwrap();
+        writer.tx().send(Ok(None)).await.unwrap();
+        writer.tx().closed().await;
     }
 
     async fn build_sender(
