+++ reductstore/src/replication/replication_task.rs
@@ -456,13 +456,17 @@ mod tests {
             let mut time = 10;
             for entry in ["test1", "test2"] {
                 for _ in 0..3 {
-                    let tx = bucket
+                    let mut writer = bucket
                         .write_record(entry, time, 4, "text/plain".to_string(), Labels::new())
                         .await
                         .unwrap();
-                    tx.send(Ok(Some(Bytes::from("test")))).await.unwrap();
-                    tx.send(Ok(None)).await.unwrap_or(());
-                    tx.closed().await;
+                    writer
+                        .tx()
+                        .send(Ok(Some(Bytes::from("test"))))
+                        .await
+                        .unwrap();
+                    writer.tx().send(Ok(None)).await.unwrap_or(());
+                    writer.tx().closed().await;
                     time += 10;
                 }
