+++ reductstore/src/storage/entry.rs
@@ -2,12 +2,14 @@
 // Licensed under the Business Source License 1.1
 
 mod entry_loader;
+mod record_writer;
 mod remove_records;
 pub(crate) mod update_labels;
+mod write_record;
 
 use crate::storage::block_manager::block_index::BlockIndex;
 use crate::storage::block_manager::{
-    spawn_read_task, spawn_write_task, BlockManager, BlockRef, RecordTx, BLOCK_INDEX_FILE,
+    spawn_read_task, BlockManager, BlockRef, RecordTx, BLOCK_INDEX_FILE,
 };
 use crate::storage::bucket::RecordReader;
 use crate::storage::entry::entry_loader::EntryLoader;
@@ -26,6 +28,10 @@ use std::sync::Arc;
 use tokio::sync::RwLock;
 use tokio::task::JoinHandle;
 use tokio::time::Instant;
+
+pub(crate) use crate::storage::entry::record_writer::{
+    RecordDrainer, RecordWriter, WriteRecordContent,
+};
 
 struct QueryHandle {
     rx: QueryRx,
@@ -78,147 +84,6 @@ impl Entry {
         options: EntrySettings,
     ) -> Result<Entry, ReductError> {
         EntryLoader::restore_entry(path, options).await
-    }
-
-    /// Starts a new record write.
-    ///
-    /// # Arguments
-    ///
-    /// * `time` - The timestamp of the record.
-    /// * `content_size` - The size of the record content.
-    /// * `content_type` - The content type of the record.
-    /// * `labels` - The labels of the record.
-    ///
-    /// # Returns
-    ///
-    /// * `Sender<Result<Bytes, ReductError>>` - The sender to send the record content in chunks.
-    /// * `HTTPError` - The error if any.
-    pub(crate) async fn begin_write(
-        &self,
-        time: u64,
-        content_size: usize,
-        content_type: String,
-        labels: Labels,
-    ) -> Result<RecordTx, ReductError> {
-        // When we write, the likely case is that we are writing the latest record
-        // in the entry. In this case, we can just append to the latest block.
-        let mut bm = self.block_manager.write().await;
-
-        let mut block_ref = if bm.index().tree().is_empty() {
-            bm.start_new_block(time, self.settings.max_block_size)
-                .await?
-        } else {
-            let block_id = *bm.index().tree().last().unwrap();
-            bm.load_block(block_id).await?
-        };
-
-        let record_type = {
-            let is_belated = {
-                let block = block_ref.write().await;
-                block.record_count() > 0 && block.latest_record_time() >= time
-            };
-            if is_belated {
-                debug!("Timestamp {} is belated. Looking for a block", time);
-                // The timestamp is belated. We need to find the proper block to write to.
-                let index_tree = bm.index().tree();
-                if *index_tree.first().unwrap() > time {
-                    // The timestamp is the earliest. We need to create a new block.
-                    debug!("Timestamp {} is the earliest. Creating a new block", time);
-                    block_ref = bm
-                        .start_new_block(time, self.settings.max_block_size)
-                        .await?;
-                    RecordType::BelatedFirst
-                } else {
-                    block_ref = bm.find_block(time).await?;
-                    let record = block_ref.read().await.get_record(time).map(|r| r.clone());
-                    // check if the record already exists
-                    if let Some(mut record) = record {
-                        // We overwrite the record if it is errored and the size is the same.
-                        return if record.state != record::State::Errored as i32
-                            || record.end - record.begin != content_size as u64
-                        {
-                            Err(ReductError::conflict(&format!(
-                                "A record with timestamp {} already exists",
-                                time
-                            )))
-                        } else {
-                            {
-                                let mut block = block_ref.write().await;
-                                record.labels = labels
-                                    .into_iter()
-                                    .map(|(name, value)| record::Label { name, value })
-                                    .collect();
-                                record.state = record::State::Started as i32;
-                                record.content_type = content_type;
-                                block.insert_or_update_record(record);
-                            }
-
-                            drop(bm); // drop the lock to avoid deadlock
-                            let tx =
-                                spawn_write_task(Arc::clone(&self.block_manager), block_ref, time)
-                                    .await?;
-                            Ok(tx)
-                        };
-                    }
-                    RecordType::Belated
-                }
-            } else {
-                // The timestamp is the latest. We can just append to the latest block.
-                RecordType::Latest
-            }
-        };
-
-        let mut block_ref = {
-            let block = block_ref.read().await;
-            // Check if the block has enough space for the record.
-            let has_no_space = block.size() + content_size as u64 > self.settings.max_block_size;
-            let has_too_many_records = block.record_count() + 1 > self.settings.max_block_records;
-
-            drop(block);
-            if record_type == RecordType::Latest && (has_no_space || has_too_many_records) {
-                // We need to create a new block.
-                debug!("Creating a new block");
-                bm.finish_block(block_ref.clone()).await?;
-                bm.start_new_block(time, self.settings.max_block_size)
-                    .await?
-            } else {
-                // We can just append to the latest block.
-                block_ref.clone()
-            }
-        };
-
-        drop(bm);
-
-        self.prepare_block_for_writing(&mut block_ref, time, content_size, content_type, labels)
-            .await?;
-
-        let tx = spawn_write_task(Arc::clone(&self.block_manager), block_ref, time).await?;
-        Ok(tx)
-    }
-
-    async fn prepare_block_for_writing(
-        &self,
-        block: &mut BlockRef,
-        time: u64,
-        content_size: usize,
-        content_type: String,
-        labels: Labels,
-    ) -> Result<(), ReductError> {
-        let mut block = block.write().await;
-        let record = Record {
-            timestamp: Some(us_to_ts(&time)),
-            begin: block.size(),
-            end: block.size() + content_size as u64,
-            content_type,
-            state: record::State::Started as i32,
-            labels: labels
-                .into_iter()
-                .map(|(name, value)| record::Label { name, value })
-                .collect(),
-        };
-
-        block.insert_or_update_record(record);
-        Ok(())
     }
 
     /// Starts a new record read.
@@ -473,278 +338,10 @@ mod tests {
         }
     }
 
-    mod begin_write {
-        use super::*;
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_write_new_block_size(path: PathBuf) {
-            let mut entry = entry(
-                EntrySettings {
-                    max_block_size: 10,
-                    max_block_records: 10000,
-                },
-                path,
-            );
-
-            write_stub_record(&mut entry, 1).await.unwrap();
-            write_stub_record(&mut entry, 2000010).await.unwrap();
-
-            let bm = entry.block_manager.read().await;
-
-            assert_eq!(
-                bm.load_block(1)
-                    .await
-                    .unwrap()
-                    .read()
-                    .await
-                    .get_record(1)
-                    .unwrap()
-                    .clone(),
-                Record {
-                    timestamp: Some(us_to_ts(&1)),
-                    begin: 0,
-                    end: 10,
-                    content_type: "text/plain".to_string(),
-                    state: record::State::Finished as i32,
-                    labels: vec![],
-                }
-            );
-
-            assert_eq!(
-                bm.load_block(2000010)
-                    .await
-                    .unwrap()
-                    .read()
-                    .await
-                    .get_record(2000010)
-                    .unwrap()
-                    .clone(),
-                Record {
-                    timestamp: Some(us_to_ts(&2000010)),
-                    begin: 0,
-                    end: 10,
-                    content_type: "text/plain".to_string(),
-                    state: record::State::Finished as i32,
-                    labels: vec![],
-                }
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_write_new_block_records(path: PathBuf) {
-            let mut entry = entry(
-                EntrySettings {
-                    max_block_size: 10000,
-                    max_block_records: 1,
-                },
-                path,
-            );
-
-            write_stub_record(&mut entry, 1).await.unwrap();
-            write_stub_record(&mut entry, 2).await.unwrap();
-            write_stub_record(&mut entry, 2000010).await.unwrap();
-
-            let bm = entry.block_manager.read().await;
-            let records = bm
-                .load_block(1)
-                .await
-                .unwrap()
-                .read()
-                .await
-                .record_index()
-                .clone();
-            assert_eq!(
-                records.get(&1).unwrap().clone(),
-                Record {
-                    timestamp: Some(us_to_ts(&1)),
-                    begin: 0,
-                    end: 10,
-                    content_type: "text/plain".to_string(),
-                    state: record::State::Finished as i32,
-                    labels: vec![],
-                }
-            );
-
-            let records = bm
-                .load_block(2000010)
-                .await
-                .unwrap()
-                .read()
-                .await
-                .record_index()
-                .clone();
-            assert_eq!(
-                records.get(&2000010).unwrap().clone(),
-                Record {
-                    timestamp: Some(us_to_ts(&2000010)),
-                    begin: 0,
-                    end: 10,
-                    content_type: "text/plain".to_string(),
-                    state: record::State::Finished as i32,
-                    labels: vec![],
-                }
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_write_belated_record(mut entry: Entry) {
-            write_stub_record(&mut entry, 1000000).await.unwrap();
-            write_stub_record(&mut entry, 3000000).await.unwrap();
-            write_stub_record(&mut entry, 2000000).await.unwrap();
-
-            let bm = entry.block_manager.read().await;
-            let records = bm
-                .load_block(1000000)
-                .await
-                .unwrap()
-                .read()
-                .await
-                .record_index()
-                .clone();
-            assert_eq!(records.len(), 3);
-            assert_eq!(
-                records.get(&1000000).unwrap().timestamp,
-                Some(us_to_ts(&1000000))
-            );
-            assert_eq!(
-                records.get(&2000000).unwrap().timestamp,
-                Some(us_to_ts(&2000000))
-            );
-            assert_eq!(
-                records.get(&3000000).unwrap().timestamp,
-                Some(us_to_ts(&3000000))
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_write_belated_first(mut entry: Entry) {
-            write_stub_record(&mut entry, 3000000).await.unwrap();
-            write_stub_record(&mut entry, 1000000).await.unwrap();
-
-            let bm = entry.block_manager.read().await;
-            let records = bm
-                .load_block(1000000)
-                .await
-                .unwrap()
-                .read()
-                .await
-                .record_index()
-                .clone();
-            assert_eq!(records.len(), 1);
-            assert_eq!(
-                records.get(&1000000).unwrap().timestamp,
-                Some(us_to_ts(&1000000))
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_write_existing_record(mut entry: Entry) {
-            write_stub_record(&mut entry, 1000000).await.unwrap();
-            write_stub_record(&mut entry, 2000000).await.unwrap();
-            let err = write_stub_record(&mut entry, 1000000).await;
-            assert_eq!(
-                err.err(),
-                Some(ReductError::conflict(
-                    "A record with timestamp 1000000 already exists"
-                ))
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_write_existing_record_belated(mut entry: Entry) {
-            write_stub_record(&mut entry, 2000000).await.unwrap();
-            write_stub_record(&mut entry, 1000000).await.unwrap();
-            let err = write_stub_record(&mut entry, 1000000).await;
-            assert_eq!(
-                err.err(),
-                Some(ReductError::conflict(
-                    "A record with timestamp 1000000 already exists"
-                ))
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_override_errored(entry: Entry) {
-            let sender = entry
-                .begin_write(1000000, 10, "text/plain".to_string(), Labels::new())
-                .await
-                .unwrap();
-
-            sender.send(Ok(None)).await.unwrap();
-            sender.closed().await;
-
-            let sender = entry
-                .begin_write(
-                    1000000,
-                    10,
-                    "text/html".to_string(),
-                    Labels::from_iter(vec![("a".to_string(), "b".to_string())]),
-                )
-                .await
-                .unwrap();
-            sender
-                .send(Ok(Some(Bytes::from(vec![0; 10]))))
-                .await
-                .unwrap();
-
-            let record = entry
-                .block_manager
-                .write()
-                .await
-                .load_block(1000000)
-                .await
-                .unwrap()
-                .read()
-                .await
-                .get_record(1000000)
-                .unwrap()
-                .clone();
-            assert_eq!(record.content_type, "text/html");
-            assert_eq!(record.labels.len(), 1);
-            assert_eq!(record.labels[0].name, "a");
-            assert_eq!(record.labels[0].value, "b");
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_begin_not_override_if_different_size(entry: Entry) {
-            let sender = entry
-                .begin_write(1000000, 10, "text/plain".to_string(), Labels::new())
-                .await
-                .unwrap();
-
-            sender.send(Ok(None)).await.unwrap();
-            sender.closed().await;
-
-            let err = entry
-                .begin_write(
-                    1000000,
-                    5,
-                    "text/html".to_string(),
-                    Labels::from_iter(vec![("a".to_string(), "b".to_string())]),
-                )
-                .await
-                .err();
-            assert_eq!(
-                err,
-                Some(ReductError::conflict(
-                    "A record with timestamp 1000000 already exists"
-                ))
-            );
-        }
-    }
-
     // Test begin_read
     mod begin_read {
         use super::*;
-        use crate::storage::storage::DEFAULT_MAX_READ_CHUNK;
+        use crate::storage::storage::MAX_IO_BUFFER_SIZE;
 
         #[rstest]
         #[tokio::test]
@@ -781,11 +378,12 @@ mod tests {
         #[rstest]
         #[tokio::test]
         async fn test_begin_read_still_written(entry: Entry) {
-            let sender = entry
+            let mut sender = entry
                 .begin_write(1000000, 10, "text/plain".to_string(), Labels::new())
                 .await
                 .unwrap();
             sender
+                .tx()
                 .send(Ok(Some(Bytes::from(vec![0; 5]))))
                 .await
                 .unwrap();
@@ -839,9 +437,9 @@ mod tests {
         #[rstest]
         #[tokio::test]
         async fn test_begin_read_ok_in_chunks(mut entry: Entry) {
-            let mut data = vec![0; DEFAULT_MAX_READ_CHUNK + 1];
+            let mut data = vec![0; MAX_IO_BUFFER_SIZE + 1];
             data[0] = 1;
-            data[DEFAULT_MAX_READ_CHUNK] = 2;
+            data[MAX_IO_BUFFER_SIZE] = 2;
 
             write_record(&mut entry, 1000000, data.clone())
                 .await
@@ -850,11 +448,11 @@ mod tests {
             let mut reader = entry.begin_read(1000000).await.unwrap();
             assert_eq!(
                 reader.rx().recv().await.unwrap().unwrap().to_vec(),
-                data[0..DEFAULT_MAX_READ_CHUNK]
+                data[0..MAX_IO_BUFFER_SIZE]
             );
             assert_eq!(
                 reader.rx().recv().await.unwrap().unwrap().to_vec(),
-                data[DEFAULT_MAX_READ_CHUNK..]
+                data[MAX_IO_BUFFER_SIZE..]
             );
             assert_eq!(reader.rx().recv().await, None);
         }
@@ -1018,11 +616,12 @@ mod tests {
         #[rstest]
         #[tokio::test]
         async fn test_entry_which_has_writer(mut entry: Entry) {
-            let sender = entry
+            let mut sender = entry
                 .begin_write(1000000, 10, "text/plain".to_string(), Labels::new())
                 .await
                 .unwrap();
             sender
+                .tx()
                 .send(Ok(Some(Bytes::from_static(b"456789"))))
                 .await
                 .unwrap();
@@ -1091,11 +690,11 @@ mod tests {
     }
 
     async fn write_record(entry: &mut Entry, time: u64, data: Vec<u8>) -> Result<(), ReductError> {
-        let sender = entry
+        let mut sender = entry
             .begin_write(time, data.len(), "text/plain".to_string(), Labels::new())
             .await?;
-        let x = sender.send(Ok(Some(Bytes::from(data)))).await;
-        sender.closed().await;
+        let x = sender.tx().send(Ok(Some(Bytes::from(data)))).await;
+        sender.tx().closed().await;
         drop(sender);
         match x {
             Ok(_) => Ok(()),
@@ -1109,11 +708,11 @@ mod tests {
         data: Vec<u8>,
         labels: Labels,
     ) -> Result<(), ReductError> {
-        let sender = entry
+        let mut sender = entry
             .begin_write(time, data.len(), "text/plain".to_string(), labels)
             .await?;
-        let x = sender.send(Ok(Some(Bytes::from(data)))).await;
-        sender.closed().await;
+        let x = sender.tx().send(Ok(Some(Bytes::from(data)))).await;
+        sender.tx().closed().await;
         match x {
             Ok(_) => Ok(()),
             Err(_) => Err(ReductError::internal_server_error("Error sending data")),
