+++ reductstore/src/storage/storage.rs
@@ -15,7 +15,7 @@ use reduct_base::error::ReductError;
 use reduct_base::msg::bucket_api::BucketSettings;
 use reduct_base::msg::server_api::{BucketInfoList, Defaults, License, ServerInfo};
 
-pub(crate) const DEFAULT_MAX_READ_CHUNK: usize = 1024 * 512;
+pub(crate) const MAX_IO_BUFFER_SIZE: usize = 1024 * 512;
 pub(crate) const CHANNEL_BUFFER_SIZE: usize = 16;
 pub(crate) const IO_OPERATION_TIMEOUT: Duration = Duration::from_secs(30);
 
@@ -274,6 +274,7 @@ mod tests {
                     .await
                     .unwrap();
                 sender
+                    .tx()
                     .send(Ok(Some(Bytes::from("0123456789"))))
                     .await
                     .unwrap();
